feat: Scans all files and downloads mistakes
Scans all the zeeko xlsx files for the 'open-ended' text field and
identifies spelling mistakes using the enchant spelling module. Each
mistake is then added to a set (to ensure no duplicates) and this set is
written to a text file, which each single word which is a spelling
mistake saved to a new line. Next step is to create logic to exclude
valid words that are not in the enchant dictionary, such as 'Facebook'
or 'Netflix'