feat: Reads valid word text file and outputs pickle set
Short Python script which opens a text file which contains words, each
one on a new line. Takes each word and adds it to a set. Outputs this
set as a pickle object. To be loaded and used when checking in
zeeko_survey_spelling.py for valid words which are indetified as
mispellings.