bug: Fixed suggested phoneme, generates candidates
Previously the suggested phoneme did not match the suggested word. This
was a dictionary overwrite error in a for loop. This has now been
corrected. When running 'ALL' for verbosity the code now outputs the
candidate lists.

Further issues:
- Code has been slowed down using 'ALL'. Need to implement
quicker candidate generation and distance calculation, not suitable for
day to day use currently.